requirements.txt
# FoodSave AI — Production Requirements
# Python 3.9+

# Deep Learning & Computer Vision
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.23.0
scikit-learn>=1.2.0
Pillow>=9.4.0

# Web Framework
Flask>=2.3.0
Flask-CORS>=4.0.0
Werkzeug>=2.3.0

# Optional: For ONNX inference
onnxruntime>=1.14.0

# Optional: For advanced routing
ortools>=9.6.2534

# Development
matplotlib>=3.7.0
seaborn>=0.12.0
tqdm>=4.65.0

# Production Server
gunicorn>=20.1.0
python-dotenv>=1.0.0
